requirements.txt
uvicorn==0.22.0
supabase==1.0.3
fastapi
python-dotenv==1.0.0
Jinja2==3.1.2
python-multipart==0.0.6